Moved getItemClass function into utilities
