--- src/items/get-item-class.js
@@ -1,8 +1,0 @@
-import { itemManifest } from "./item-manifest.js";
-
-export var itemClasses = {};
-export async function getItemClass(itemName, type, scene) {
-    let itemClass = itemClasses[type + itemName];
-
-    return new itemClass.default(scene);
-}
