+++ src/items/item.js
@@ -1,7 +1,7 @@
 import { ClickableObject } from "../clickable-object.js";
 import { OBJECT_TYPES, CONSTANTS } from "../constants/constants.js";
 import { itemManifest } from "./item-manifest.js";
-import { getItemClass } from "./get-item-class.js";
+import { getItemClass } from "../utilities.js";
 
 export class Item extends ClickableObject {
     // Text data
